socialnetwork/requirements.txt
Django>=5.0.2,<5.1
psycopg2-binary>=2.9.9,<3.0
python-decouple>=3.8,<3.9
drf-yasg>=1.21.7
djangorestframework>=3.14.0,<3.15
djangorestframework-simplejwt>=5.3.1,<5.4
pillow>=10.2.0,<10.3
django-allauth >= 0.61.0,<0.62.0
django-crispy-forms >= 2.1,<2.2
channels>=4.0.0,<4.1
daphne>=4.1.0,<4.2
channels-redis>=4.2.0,<4.3
aioredis>=2.0.1,<2.1
asyncio>=3.4.3,<3.5
django-storages>=1.14.2,<1.15
django-minio-storage>=0.5.7
django-minio-backend>=3.1.0,<3.2
django-cors-headers>=4.3.1,<4.4
cryptography>=42.0.3,<42.1.0
python-dotenv>=1.0.1,<1.1.0
factory_boy>=3.3.0,<3.4.0